+++ src/Traits/CreatesFromArrayOrEntity.php
@@ -10,7 +10,9 @@ use Nandan108\DtoToolkit\Contracts\ScopedPropertyAccessInterface;
 use Nandan108\DtoToolkit\Core\ProcessingErrorList;
 use Nandan108\DtoToolkit\Enum\ErrorMode;
 use Nandan108\DtoToolkit\Enum\Phase;
+use Nandan108\DtoToolkit\Enum\PresencePolicy;
 use Nandan108\DtoToolkit\Exception\Config\InvalidConfigException;
+use Nandan108\DtoToolkit\Exception\Process\ExtractionException;
 use Nandan108\PropAccess\PropAccess;
 
 /**
@@ -24,7 +26,7 @@ use Nandan108\PropAccess\PropAccess;
 trait CreatesFromArrayOrEntity
 {
     /**
-     * Create a new instance of the DTO from a request.
+     * Create a new instance of the DTO from an array.
      *
      * @param bool $ignoreUnknownProps If true, unknown properties will be ignored
      *
@@ -36,7 +38,8 @@ trait CreatesFromArrayOrEntity
         array $input,
         bool $ignoreUnknownProps = false,
         ?ProcessingErrorList $errorList = null,
-        ?ErrorMode $errorMode = null,   // <── yes
+        ?ErrorMode $errorMode = null,
+        bool $clear = true,
     ): static {
         // fill the DTO with the input values
         /** @psalm-suppress InaccessibleMethod */
@@ -55,18 +58,58 @@ trait CreatesFromArrayOrEntity
             $toBeFilled = array_intersect($fillables, $propsInScope);
         }
 
+        $presencePolicy = static::getPresencePolicy();
+        $defaultValues = static::getDefaultValues();
+
         // get mappers for the properties to be filled
         $mappers = MapFrom::getMappers($this, $toBeFilled);
 
         $inputToBeFilled = array_intersect_key($input, array_flip($toBeFilled));
 
+        // PresencePolicy determine which props are marked as "filled" :
+        // Default:             any prop with a value present in input, including NULL values.
+        // NullMeansMissing:    Same as Default, NULL values excluded (treated as missing)
+        // MissingMeansDefault: Always marked as filled, even if missing in input (gets default value)
+
         // Get data from mappers. Only fill if we actually get a value.
-        foreach ($mappers as $prop => $mapper) {
-            $inputToBeFilled[$prop] = $mapper($input, $this);
+        foreach ($presencePolicy as $prop => $policy) {
+            if ($mapper = $mappers[$prop] ?? null) {
+                try {
+                    // get value from mapper
+                    $inputToBeFilled[$prop] = $mapper($input, $this);
+                } catch (ExtractionException $e) {
+                    // Currently, extraction failures will be treated as missing input, and #[MapFrom(path)] fails
+                    // immediately (by default) on missing input, even as part of a multi-path expression.
+                    // This can be adjusted via ThrowMode passed to MapFrom attribute: #[MapFrom(paths, ThrowMode::NEVER)],
+                    // but in this case, we lose distinction between "missing input" vs. "null input", which is not ideal.
+                    // TODO: refine prop-path exception handling, so that we can
+                    // distinguish between "missing input" vs. "null input" and handle accordingly.
+                    unset($inputToBeFilled[$prop]);
+                    continue;
+                }
+            }
+            // handle MissingMeansDefault policy
+            if (!array_key_exists($prop, $inputToBeFilled)) {
+                if (PresencePolicy::MissingMeansDefault === $policy) {
+                    $inputToBeFilled[$prop] = $defaultValues[$prop];
+                    continue;
+                }
+            }
+            // handle NullMeansMissing policy
+            elseif (null === $inputToBeFilled[$prop]) {
+                if (PresencePolicy::NullMeansMissing === $policy) {
+                    unset($inputToBeFilled[$prop]);
+                }
+            }
         }
 
         // and fill the DTO
         $this->fill($inputToBeFilled);
+
+        // reset unfilled properties to default values
+        if ($clear) {
+            $this->clear(excludedPropsMap: $inputToBeFilled);
+        }
 
         // cast the values to their respective types and return the DTO
         if ($this instanceof ProcessesInterface) {
@@ -86,15 +129,16 @@ trait CreatesFromArrayOrEntity
      *
      * @throws InvalidConfigException
      *
-     * @psalm-suppress PossiblyUnusedMethod, MethodSignatureMismatch
+     * @psalm-suppress PossiblyUnusedMethod, PossiblyUnusedReturnValue, MethodSignatureMismatch
      */
     public function _fromArrayLoose(
         array $input,
         ?ProcessingErrorList $errorList = null,
-        ?ErrorMode $errorMode = null,   // <── yes
+        ?ErrorMode $errorMode = null,
+        bool $clear = true,
     ): static {
         /** @psalm-suppress NoValue */
-        return $this->_fromArray($input, ignoreUnknownProps: true, errorList: $errorList, errorMode: $errorMode);
+        return $this->_fromArray($input, true, $errorList, $errorMode, $clear);
     }
 
     /**
@@ -104,7 +148,8 @@ trait CreatesFromArrayOrEntity
         object $entity,
         bool $ignoreInaccessibleProps = true,
         ?ProcessingErrorList $errorList = null,
-        ?ErrorMode $errorMode = null,   // <── yes
+        ?ErrorMode $errorMode = null,
+        bool $clear = true,
     ): static {
         /** @var array */
         $inputData = PropAccess::getValueMap(
@@ -113,6 +158,6 @@ trait CreatesFromArrayOrEntity
             ignoreInaccessibleProps: $ignoreInaccessibleProps,
         );
 
-        return $this->_fromArray($inputData, true, $errorList, $errorMode);
+        return $this->_fromArray($inputData, true, $errorList, $errorMode, $clear);
     }
 }