+++ tests/Unit/Casting/MappingTest.php
@@ -9,9 +9,13 @@ use Nandan108\DtoToolkit\Attribute\MapTo;
 use Nandan108\DtoToolkit\Attribute\Outbound;
 use Nandan108\DtoToolkit\Core\FullDto;
 use Nandan108\DtoToolkit\Exception\Config\InvalidConfigException;
-use Nandan108\DtoToolkit\Exception\Process\ExtractionException;
+use Nandan108\DtoToolkit\Exception\Process\GuardException;
 use Nandan108\DtoToolkit\Tests\Traits\CanTestCasterClassesAndMethods;
+use Nandan108\DtoToolkit\Validate;
+use Nandan108\PropPath\Support\ThrowMode;
 use PHPUnit\Framework\TestCase;
+
+use function PHPUnit\Framework\never;
 
 final class MappingTest extends TestCase
 {
@@ -20,7 +24,7 @@ final class MappingTest extends TestCase
     public function testMapFromExistingInput(): void
     {
         $dtoClass = new class extends FullDto {
-            #[MapFrom(['qux1' => 'foo', 'qux2' => ['baz', 'bar']])]
+            #[MapFrom(['qux1' => 'foo', 'qux2' => ['baz', 'bar']], ThrowMode::NEVER)]
             public string | array | null $bar = '';
 
             #[MapFrom('bar')]
@@ -49,6 +53,8 @@ final class MappingTest extends TestCase
         $dto = $dtoClass::fromArrayLoose([
             'foo' => 'FOO-val',
             'bar' => 'BAR-val',
+            // 'baz' is missing, so bar prop's mapper fails.
+            // But since we used ThrowMode::NEVER, it just returns a null value.
         ]);
         $this->assertSame(['qux1' => 'FOO-val', 'qux2' => [null, 'BAR-val']], $dto->bar);
     }
@@ -56,39 +62,48 @@ final class MappingTest extends TestCase
     public function testMapFromMissingInputThrowsIfLooseRequiredByOneBang(): void
     {
         $dtoClass = new class extends FullDto {
-            #[MapFrom(['qux1' => 'foo', 'qux2' => ['!baz', 'bar']])]
+            #[MapFrom(['qux1' => 'foo', 'qux2' => ['!baz', 'bar']], ThrowMode::NEVER)]
             public string | array | null $bar = '';
         };
 
         $dto = $dtoClass::fromArrayLoose([
             'foo' => 'FOO-val',
             'bar' => 'BAR-val',
-            // null value doesn't throw
+            // default throw mode = ThrowMode::NEVER -> null value doesn't throw
             'baz' => null,
         ]);
         $this->assertSame(['qux1' => 'FOO-val', 'qux2' => [null, 'BAR-val']], $dto->bar);
 
-        $this->expectException(ExtractionException::class);
         /** @psalm-suppress UnusedVariable */
-        $dtoClass::fromArrayLoose([
+        $dto->clear()->_fromArrayLoose([
             'foo' => 'FOO-val',
             'bar' => 'BAR-val',
+            // default throw mode = ThrowMode::NEVER, but overriden by bang ('!baz')
+            // -> missing value makes the mapper fail and throw.
+            // Mapper throws are converted treated as "missing value", so bar prop
+            // is kept untoutched and keeps its default value (empty string)
         ]);
+        $this->assertSame(false, $dto->isFilled('bar'));
+        $this->assertSame($dto->getDefaultValues()['bar'], $dto->bar);
     }
 
     public function testMapFromNullInputThrowsIfStrictRequiredByTwoBangs(): void
     {
-        $dtoClass = new class extends FullDto {
+        $dtoClassWithMapperThatThrows = new class extends FullDto {
             #[MapFrom(['qux1' => 'foo', 'qux2' => ['!!baz', 'bar']])]
-            public string | array | null $bar = '';
+            #[Validate\NotNull] // guard not executed
+            public ?array $bar = null;
         };
 
-        $this->expectException(ExtractionException::class);
-        $dtoClass::fromArrayLoose([
+        $dto = $dtoClassWithMapperThatThrows::fromArrayLoose([
             'foo' => 'FOO-val',
             'bar' => 'BAR-val',
-            'baz' => null, // null value will throw in this case (!!baz)
+            'baz' => null, // null value makes mapper fail in this case (!!baz), resulting in missing value
         ]);
+        // Since the bar prop is not marked as filled, it is not further processed by the NotNull validator,
+        // therefore no GuardException is thrown here.
+
+        $this->assertSame(false, $dto->isFilled('bar'));
     }
 
     public function testMapFromThrowsIfUsedInOutboundPhase(): void